Record AV qtype: start video capture before starting recording #460009
This gives the camera time to sort out its autofocus and white balance.

It also gives the user time to verify they are happy with what is being
recorded, before recording starts for real.

This is achieved by having an additional state for the button for video
inputs. Now the button starts saying 'Start camera' after that is
clicked, the camera is started with the preview, and the button changes
to 'Start recording'. Once that is clicked, then things are as before.

There is no change for audio inputs.