Record AV question type: option for per-input feedback #467223
For questions with mulitple inputs (Customised audio/video) there is
now the option to input some feedback to be shown next to each
recording widget, once the question has been submitted.

This only shows if a recording has been made in the corresponding input.

The feedback data is stored in the question_answers table, allowing
us to reuse some of the core Moodle handling of this data.