RecordAV/Combined: put example placeholder in text inputs #513612
This is on the editing form for these question types. It ensures that
when you copy them, to paste into the question text, they come as plain
text, without unwanted HTML styling.

There is also some JavaScript to automatically select all when you
click on them.

Finally, the label of the 'Update the form' button was changed.