Record A/V: improve form validation and auto-fix some errors #534748
The validation now finds all errors in all placehoders at the same time.

It can also fix up some of those errors automatically. When this is done,
the form is still redisplayed with the validation message, to help
users learn how to get it right first time.

Also, the validation is now run whne the 'Update form' button is
pressed.

To facilitate this, if there are no errors when the button is clicked,
the message 'The form was updated. You can add per-input feedback below.'
is displayed by the button.

To help test this, the validate_widget_placeholders unit tests
now use a data provider.